Wavelength dithers decimal nm

+++ modules/core/src/main/scala/lucuma/odb/api/model/gmos/longslit/DeltaWavelengthCalculator.scala
@@ -9,7 +9,7 @@ import lucuma.core.math.units.Nanometer
 
 trait DeltaWavelengthCalculator[G] {
 
-  def Δλ(g: G): Quantity[Int, Nanometer]
+  def Δλ(g: G): Quantity[BigDecimal, Nanometer]
 
 }
